requirements.txt
pytecplot==1.6.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
matplotlib-inline==0.1.7